requirements.txt
Pillow
python-dotenv
redis
numpy
scipy
pytest